move teams from info to actions
